Move generate_kfrags() to pre.rs
