+++ umbral-pre/src/lib.rs
@@ -107,7 +107,9 @@ mod traits;
 pub use capsule::{Capsule, OpenReencryptedError};
 pub use capsule_frag::CapsuleFrag;
 pub use dem::{DecryptionError, EncryptionError};
-pub use key_frag::{generate_kfrags, KeyFrag};
+pub use key_frag::KeyFrag;
 pub use keys::{PublicKey, SecretKey, SecretKeyFactory, SecretKeyFactoryError, Signature, Signer};
-pub use pre::{decrypt_original, decrypt_reencrypted, encrypt, reencrypt, ReencryptionError};
+pub use pre::{
+    decrypt_original, decrypt_reencrypted, encrypt, generate_kfrags, reencrypt, ReencryptionError,
+};
 pub use traits::{DeserializationError, SerializableToArray};
